Add cpp class for tpch_grammer
